requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# Database
psycopg2-binary>=2.9,<3.0

# Authentication & Permissions
djangorestframework-simplejwt>=5.3,<6.0
django-guardian>=2.4,<3.0

# Workflow (custom implementation, no external dependency needed)

# Multi-tenancy
django-tenants>=3.6,<4.0

# Async Tasks
celery>=5.3,<6.0
redis>=5.0,<6.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Utilities
python-decouple>=3.8,<4.0
django-extensions>=3.2,<4.0
Pillow>=10.0,<11.0

# Export
openpyxl>=3.1,<4.0
reportlab>=4.0,<5.0

# Development
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
factory-boy>=3.3,<4.0
black>=23.0
flake8>=6.0
